+++ biopipen/ns/scrna.py
@@ -3232,8 +3232,10 @@ class CellCellCommunicationPlots(Proc):
             By default, the last column will be used. If the method doesn't have a specificity, set it to None.
         ligand_expr: The name of the column containing the mean (or otherwise summarized) expression of the ligand.
             Only used when plot_type = "linkedheatmap".
+            If not specified, `ligand_means` will be tried first, then `ligand_trimean`.
         receptor_expr: The name of the column containing the mean (or otherwise summarized) expression of the receptor.
             Only used when plot_type = "linkedheatmap".
+            If not specified, `receptor_means` will be tried first, then `receptor_trimean`.
         devpars (ns): The parameters for the plot.
             - res (type=int): The resolution of the plot
             - height (type=int): The height of the plot
@@ -3258,8 +3260,8 @@ class CellCellCommunicationPlots(Proc):
         "subset": None,
         "magnitude": None,
         "specificity": None,
-        "ligand_expr": "ligand_means",
-        "receptor_expr": "receptor_means",
+        "ligand_expr": None,
+        "receptor_expr": None,
         "devpars": {"res": 100},
         "more_formats": [],
         "descr": "Cell-cell communication plot",
@@ -3379,6 +3381,8 @@ class Slingshot(Proc):
         dims (type=auto): The dimensions to use for the analysis.
             A list or a string with comma separated values.
             Consecutive numbers can be specified with a colon (`:`) or a dash (`-`).
+            Or a single number greater than 1, which will be expanded to `1:number`.
+            If `None`, all dimensions will be used.
         start: The starting group for the Slingshot analysis.
         end: The ending group for the Slingshot analysis.
         prefix: The prefix to add to the column names of the resulting pseudotime variable.
@@ -3393,7 +3397,7 @@ class Slingshot(Proc):
     envs = {
         "group_by": None,
         "reduction": None,
-        "dims": [1, 2],
+        "dims": None,
         "start": None,
         "end": None,
         "prefix": None,